% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{chg2pct}
\alias{chg2pct}
\title{Convert last two values in a series and convert them to a percent
increase or decrease}
\usage{
chg2pct(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
The percentage change over the last two time periods.
}
\description{
Takes a column named 'date' and column named 'value' of a dataframe,
arranges the values by date, and computes the percentage change.
}
\details{
dataframe must have at least two columns: one named 'date' and the
  other named 'value'.
}
\examples{
\dontrun{
chg2pct(pos_data_set)
chg2pct(neg_data_set)
# fail
chg2pct("a")
chg2pct(data.frame(a = c(1, 2)))
chg2pct(data.frame(a = 1, b = 1))
chg2pct(data.frame(a = c(1, 2), b = c(1, 2)))
}
}
