% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{color_box}
\alias{color_box}
\title{Return a color for valueBox}
\usage{
color_box(number = 0)
}
\arguments{
\item{number}{a number}
}
\value{
a color for valueBox
}
\description{
Takes a value and returns the color 'green' if positive, the
color 'red', if negative, and the color 'blue' for anything else.
}
\details{
For the arguments for valueBox, query `?shinydashboard::valueBox`.
}
\examples{
\dontrun{
color_box(number = 0)
color_box(number = 1)
color_box(number = -1)
color_box(number = NA)
}
}
