% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_csushpinsa.R
\docType{data}
\name{csushpinsa}
\alias{csushpinsa}
\title{Case-Shiller U.S. National Home Price Index.}
\format{
A dataframe where each row represents a month and 2 variables:

\describe{
  \item{date}{date the index was computed}
  \item{value}{index value for single-family home prices}
}
}
\source{
\url{https://www.spglobal.com/spdji/en/indices/indicators/sp-corelogic-case-shiller-us-national-home-price-nsa-index/#overview}
}
\usage{
csushpinsa
}
\description{
A dataset containing the monthly economic index for the U.S.
}
\details{
According to the S & P Dow Jones Indices, "the S&P CoreLogic
Case-Shiller U.S. National Home Price NSA Index is a composite of
single-family home price indices for the nine U.S. Census divisions
and is calculated monthly." The index "seeks to measure changes
in the total value of all existing single-family housing stock."
}
\keyword{datasets}
