% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{insert_arrow}
\alias{insert_arrow}
\title{Return an arrow icon for valueBox}
\usage{
insert_arrow(number = 0)
}
\arguments{
\item{number}{a number}
}
\value{
arrow icon for valueBox
}
\description{
Takes a value and returns an up arrow if positive, a down arrow if
negative, and a horizontal arrows for zero.
}
\details{
For the arguments for valueBox, query `?shinydashboard::valueBox`.
}
\examples{
\dontrun{
arrow_direction(number = 0)
arrow_direction(number = 1)
arrow_direction(number = -1)
arrow_direction(number = NA)
}

}
