% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_umcsent.R
\docType{data}
\name{umcsent}
\alias{umcsent}
\title{University of Michigan Consumer Sentiment Survey.}
\format{
A dataframe where each row represents a month and 2 variables:

\describe{
  \item{date}{date the survey was collected}
  \item{value}{index of consumer sentiment}
}
}
\source{
\url{https://data.sca.isr.umich.edu/fetchdoc.php?docid=24774}
}
\usage{
umcsent
}
\description{
A dataset containing the average monthly sentiment of U.S. consumers
}
\details{
The Index of Consumer Expectations, published by the University
of Michigan, focuses on three areas: how
consumers view prospects for their own financial situation, how they
view prospects for the general economy over the near term, and their
view of prospects for the economy over the long term.
}
\keyword{datasets}
