% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cycles.R
\docType{data}
\name{us_recessions}
\alias{us_recessions}
\title{US business us_recessions.}
\format{
A data frame where each row represents the  and 5 variables:

\describe{
  \item{peak}{date economic activity peaked}
  \item{trough}{date economic activity reached its lowest point}
}
}
\source{
\url{https://www.nber.org/research/business-cycle-dating/business-cycle-dating-procedure-frequently-asked-questions}
}
\usage{
us_recessions
}
\description{
A dataset containing the US business cycle expansions and contractions from 1857 forward
}
\details{
The NBER’s Business Cycle Dating Committee maintains a chronology
of US business us_recessions. The chronology identifies the dates of peak and
trough months in economic activity. The peak is the month in which a
variety of economic indicators reach their highest level, followed by
a significant decline in economic activity. Similarly, a month is
designated as a trough when economic activity reaches a low point and
begins to rise again for a sustained period.
}
\keyword{datasets}
