% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wei.R
\docType{data}
\name{wei}
\alias{wei}
\title{Weekly Economic indicators index.}
\format{
A time series object where each row represents a week and 2 variables:

\describe{
  \item{date}{date the index was computed}
  \item{value}{index value}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
wei
}
\description{
A dataset containing the weekly economic index for the U.S.
}
\details{
The WEI is an index of real economic activity using timely and
relevant high-frequency data. It represents the common component of ten
different daily and weekly series covering consumer behavior, the labor
market, and production. The WEI is scaled to the four-quarter GDP growth
rate; for example, if the WEI reads -2 percent and the current level of
the WEI persists for an entire quarter, we would expect, on average, GDP
that quarter to be 2 percent lower than a year previously.
}
\keyword{datasets}
